requirements.txt
pandas
sentence-transformers
faiss-cpu
fastapi
uvicorn
python-multipart
sqlalchemy
alembic
psycopg2-binary
pydantic
rq
redis
huggingface_hub[hf_xet]
matplotlib
scikit-learn
numpy
umap-learn
